pyproject.toml

[project]
# PEP 621 project metadata
# See https://www.python.org/dev/peps/pep-0621/
dependencies = ["aiohttp==3.8.3", "aiosignal==1.3.1", "async-timeout==4.0.2", "attrs==22.2.0", "charset-normalizer==2.1.1", "colorama==0.4.6", "discord-py==2.1.0", "frozenlist==1.3.3", "idna==3.4", "iniconfig==2.0.0", "multidict==6.0.4", "packaging==23.0", "pluggy==1.0.0", "pytest==7.2.1", "python-dotenv==0.21.0", "Unidecode==1.3.6", "yarl==1.8.2"]
requires-python = ">=3.11"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
